requirements-gpu.txt
absl-py==0.2.2
astor==0.7.1
gast==0.2.0
grpcio==1.13.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
protobuf==3.6.0
six==1.11.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1
